edit shipment repository

+++ src/main/java/com/example/cms/model/repository/ShipmentRepository.java
@@ -18,17 +18,17 @@ public interface ShipmentRepository extends JpaRepository<Shipment, Integer> {
     //get shipment date
     @Query(value = "select shipmentMonth, shipmentDay, shipmentYear from Shipment s where " +
             "shipmentId = :shipmentId", nativeQuery = true)
-    Shipment retrieveDateShipment(@Param("shipmentId") long shipmentId);
+    Shipment retrieveDateShipment(@Param("shipmentId") int shipmentId);
 
     //get shipment size
     @Query(value = "select size from Shipment s where " +
             "shipmentId = :shipmentId", nativeQuery = true)
-    Shipment retrieveSize(@Param("shipmentId") long shipmentId);
+    Shipment retrieveSize(@Param("shipmentId") int shipmentId);
 
     // get new lot number for the new products that are shipping
     @Query(value = "select newLotNumber from Shipment s where " +
             "shipmentId = :shipmentId", nativeQuery = true)
-    Shipment retrieveNewLotNum(@Param("shipmentId") long shipmentId);
+    Shipment retrieveNewLotNum(@Param("shipmentId") int shipmentId);
 
     // get ALL new itemId for the new products that are shipping
     @Query(value = "select newItemId from Shipment s where " +
@@ -38,15 +38,15 @@ public interface ShipmentRepository extends JpaRepository<Shipment, Integer> {
     // get new expiry date for new products
     @Query(value = "select newExpiryMonth, newExpiryDay, newExpiryYear from Shipment s where " +
             "shipmentId = :shipmentId", nativeQuery = true)
-    Shipment retrieveNewExpiryDate(@Param("shipmentId") long shipmentId);
+    Shipment retrieveNewExpiryDate(@Param("shipmentId") int shipmentId);
 
     // get status of new products that are shipping
     @Query(value = "select shipmentStatus from Shipment s where " +
             "shipmentId = :shipmentId", nativeQuery = true)
-    Shipment retrieveStatus(@Param("shipmentId") long shipmentId);
+    Shipment retrieveStatus(@Param("shipmentId") int shipmentId);
 
     // get reason for cancellation
     @Query(value = "select cancellationReason from Shipment s where " +
             "shipmentId = :shipmentId", nativeQuery = true)
-    Shipment retrieveReason(@Param("shipmentId") long shipmentId);
+    Shipment retrieveReason(@Param("shipmentId") int shipmentId);
 }
